requirements.txt
django
djangorestframework
djangorestframework-jwt
cryptography
pyjwt
python-dotenv
psycopg2
psycopg2-binary
social-auth-app-django
requests
request
